requirements.txt
# Core ML and Data Science
numpy>=1.26.0,<2.0.0
pandas>=2.0.0,<3.0.0
scikit-learn>=1.5.0,<2.0.0
scipy>=1.11.0

# Network Analysis
networkx>=3.1,<4.0

# Graph2Vec - Custom implementation using gensim
gensim>=4.0.0

# Visualization
matplotlib>=3.7.0,<4.0.0
seaborn>=0.12.0,<1.0.0

# Data Collection
kaggle>=1.5.16
requests>=2.31.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Jupyter Notebooks
jupyter>=1.0.0
ipykernel>=6.25.0

# Utilities
tqdm>=4.65.0
joblib>=1.3.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
